utilty.py updated
