+++ utils/utility.py
@@ -1,3 +1,4 @@
+import pandas as pd
 
 """
 @PARAMS
@@ -99,3 +100,44 @@ def agg_categorical_column(series):
 def agg_numerical_column(series):
     return series.mean()
 
+
+def anonymizer(df, partitions, feature_columns, sensitive_column, categorical, max_partitions=None):
+    aggregations = {}
+
+    for column in feature_columns:
+        if column in categorical:
+            aggregations[column] = agg_categorical_column
+        else:
+            aggregations[column] = agg_numerical_column
+    rows = []
+
+    for i, partition in enumerate(partitions):
+        if i % 100 == 1:
+            print("Finished {} partitions.".format(i))
+        if max_partitions is not None and i > max_partitions:
+            break
+        grouped_columns = df.loc[partition].assign(
+            _common88column_=1).groupby('_common88column_').agg(aggregations, squeeze=False)
+        sensitive_counts = df.loc[partition].groupby(
+            sensitive_column).agg({sensitive_column: 'count'})
+        values = grouped_columns.iloc[0].to_dict()
+        for sensitive_value, count in sensitive_counts[sensitive_column].items():
+            if count == 0:
+                continue
+            values.update({
+                sensitive_column: sensitive_value,
+                'count': count,
+
+            })
+            rows.append(values.copy())
+    dfn = pd.DataFrame(rows)
+    pdfn = dfn.sort_values(feature_columns+[sensitive_column])
+    return pdfn
+
+
+def get_full_span(df, categorical):
+    for name in df.columns:
+        if name not in categorical:
+            df[name] = pd.to_numeric(df[name])
+
+    return get_spans(df, categorical, df.index)
